% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongodb_tools.r
\name{mongodb_connection}
\alias{mongodb_connection}
\title{Create a MongoDB connection}
\usage{
mongodb_connection(
  host,
  user,
  pass,
  db = "cybersecurity_articles",
  collection = "metadata"
)
}
\arguments{
\item{host}{MongoDB host name or IP.}

\item{user}{Username for authentication.}

\item{pass}{Password for authentication.}

\item{db}{Database name (default: "cybersecurity_articles").}

\item{collection}{Collection name (default: "metadata").}
}
\value{
A mongolite::mongo connection object.
}
\description{
Create a MongoDB connection
}
