% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongodb_tools.r
\name{mongodb_load_json}
\alias{mongodb_load_json}
\title{Load and insert JSON data into MongoDB}
\usage{
mongodb_load_json(file_path, collection_obj)
}
\arguments{
\item{file_path}{Path to JSON file.}

\item{collection_obj}{A mongolite::mongo collection connection.}
}
\value{
Invisibly returns number of inserted records.
}
\description{
Load and insert JSON data into MongoDB
}
