% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_oai_records.r
\name{parse_oai_records}
\alias{parse_oai_records}
\title{Parse OAI-PMH Records}
\usage{
parse_oai_records(xml_content)
}
\arguments{
\item{xml_content}{XML document from arXiv OAI-PMH result}
}
\value{
A list of parsed records
}
\description{
Parse OAI-PMH Records
}
\keyword{internal}
